Replace PsiClassHelper.kt
